clean carla_walker_agent
