--- carla_walker_agent/src/carla_walker_agent/misc.py
@@ -1,70 +1,0 @@
-#!/usr/bin/env python
-#
-# Copyright (c) 2018-2020 Intel Corporation
-#
-# This work is licensed under the terms of the MIT license.
-# For a copy, see <https://opensource.org/licenses/MIT>.
-#
-""" This module contains helper functions """
-
-import math
-import numpy as np
-
-
-def is_within_distance_ahead(target_location, current_location, orientation, max_distance):
-    """
-    Check if a target object is within a certain distance in front of a reference object.
-
-    :param target_location: location of the target object
-    :param current_location: location of the reference object
-    :param orientation: orientation of the reference object
-    :param max_distance: maximum allowed distance
-    :return: True if target object is within max_distance ahead of the reference object
-    """
-    target_vector = np.array([target_location.x - current_location.x,
-                              target_location.y - current_location.y])
-    norm_target = np.linalg.norm(target_vector)
-
-    # If the vector is too short, we can simply stop here
-    if norm_target < 0.001:
-        return True
-
-    if norm_target > max_distance:
-        return False
-
-    forward_vector = np.array(
-        [math.cos(math.radians(orientation)), math.sin(math.radians(orientation))])
-    d_angle = math.degrees(math.acos(np.dot(forward_vector, target_vector) / norm_target))
-
-    return d_angle < 90.0
-
-
-def compute_magnitude_angle(target_location, current_location, orientation):
-    """
-    Compute relative angle and distance between a target_location and a current_location
-
-    :param target_location: location of the target object
-    :param current_location: location of the reference object
-    :param orientation: orientation of the reference object
-    :return: a tuple composed by the distance to the object and the angle between both objects
-    """
-    target_vector = np.array([target_location.x - current_location.x,
-                              target_location.y - current_location.y])
-    norm_target = np.linalg.norm(target_vector)
-
-    forward_vector = np.array([math.cos(math.radians(orientation)),
-                               math.sin(math.radians(orientation))])
-    d_angle = math.degrees(
-        math.acos(np.clip(np.dot(forward_vector, target_vector) / norm_target, -1., 1.)))
-
-    return (norm_target, d_angle)
-
-
-def distance_vehicle(pos1, pos2):
-    """
-    calculate distance between waypoint and vehicle position
-    """
-    dx = pos1.x - pos2.x
-    dy = pos1.y - pos2.y
-
-    return math.sqrt(dx * dx + dy * dy)
